feat(core): autoconvert html to md when pasting in markdown mode
